+++ src/domain/ports/input_cache.rs
@@ -16,7 +16,7 @@ pub enum InputCacheError {
 }
 
 pub trait InputCache {
-    fn save(input: &str, year: u16, day: u8) -> Result<(), InputCacheError>;
-    fn load(year: u16, day: u8) -> Result<String, InputCacheError>;
+    fn save(input: &str, year: i32, day: i32) -> Result<(), InputCacheError>;
+    fn load(year: i32, day: i32) -> Result<String, InputCacheError>;
     fn clear() -> Result<(), InputCacheError>;
 }
